Hit /refresh_token on session start
This way, a session is started on the client by returning a jwt
that will be stored in memory. This requires the cookie
recipes_refresh_token to be present when the api request is made.

Then, all other api requests are enabled via axios since a signal
sessionStarted is set to true.